backend/pyproject.toml
[tool.poetry]
name = "turbocommerce"
version = "0.1.0"
description = ""
authors = ["ashdaily <ashtokyo31@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.2"
django-rest-framework-social-oauth2 = "^1.1.0"
djangorestframework = "^3.12.1"
django-environ = "^0.4.5"
psycopg2 = "^2.8.6"
django-cors-headers = "^3.5.0"
pyhustler = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
